generalized dftw to encompass q codelets.  As a side effect, q codelets are now threaded. [empty commit message]

+++ libbench2/my-getopt.c
@@ -30,12 +30,12 @@ static const char *scan_pointer = 0;
 void my_usage(const char *progname, const struct my_option *opt)
 {
     int i;
-    int col = 0;
+    size_t col = 0;
 
     fprintf(stdout, "Usage: %s", progname);
     col += (strlen(progname) + 7);
     for (i = 0; opt[i].long_name; i++) {
-	int option_len;
+	size_t option_len;
 
 	option_len = strlen(opt[i].long_name);
 	if (col >= 80 - (option_len + 16)) {
@@ -99,7 +99,7 @@ int my_getopt(int argc, char *argv[], const struct my_option *optarray)
 	  ++p;
 	  
 	  for (l = optarray; l->short_name; ++l) {
-	       int len = strlen(l->long_name);
+	       size_t len = strlen(l->long_name);
 	       if (!strncmp(l->long_name, p, len) && 
 		   (!p[len] || p[len] == '=')) {
 		    switch (l->argtype) {
